Refactoring: create separate class for sort array
