+++ easy-coding-standard.yml
@@ -13,7 +13,7 @@ parameters:
 
         SlevomatCodingStandard\Sniffs\Classes\UnusedPrivateElementsSniff.UnusedMethod:
             - '*/src/Model/YamlInline/YamlInlineChecker.php'
-            - '*/src/Model/YamlAlphabetical/YamlAlphabeticalDataFactory.php'
+            - '*/src/Model/YamlAlphabetical/YamlSortService.php'
 
         PhpCsFixer\Fixer\Whitespace\SingleBlankLineAtEofFixer:
             - '*/tests/Model/YamlEmptyLineAtEnd/resource/unfixed/foo.php' # this is only for unit test so this has to be skipped
